module_1_DLTSInterface/protocol_Ethereum/ethereum-parser/requirements.txt
ijson==3.2.3
datrie==0.8.2
python-dotenv==1.0.1
py-evm==0.8.0b1
